Reorganised folder structure for better version control
Instead of having just one version of the plugin sitting in the root of the repository I decided to create a separate folder for each Bludit version in which to keep the actual plugin.
Apart from helping to identify, which Bludit version is compatible with which plugin version it gets around the Github habit of adding a suffix of "master" to the end of the download folder.